// frontend/src/types/Car/Car.ts
import { StringLiteral } from "typescript";

interface Car {
    Car_Id: number;
    FuelType: string;
    State: string;
    Vehicle_Year: number;
    Number_of_trips: number;
    Vehicle_Type: string;
    Vehicle_Make: String;
    Vehicle_Model: String;
    Daily_Price: number
}

export default Car;